Update patch set 1
Patch Set 1:

in general, before we submit this code (and until we get the integration tests up and running automatically) we must run the application and test our features manually.

bare in mind that once we have automatic integration tests (junit classes that loads & runs the ODL with our code) we'll be able to add automatic tests that verifies features like the LISP, REST API, etc. so adding such tests should probably be part of the commit itself - thoughts?

Patch-set: 1
Label: Verified=0