Update patch set 3
Patch Set 3:

(2 comments)

Thanks for the feedback Guy. I've just submitted a patch covering both your comments 1) and 2). Take a look at the code and let me know what you think. 

Regarding your comment 3), I agree that it's a dirty solution. The problem is that the XML/JSON annotated classes must contain the fields to be used on XML/JSON, and it's not possible to use the "attribute-less" LispAddress class for that. We agreed on the workaround of a LispAddressGeneric class after some discussion with Gal. I'm not happy with that either, so please feel free to propose alternatives ;)

Patch-set: 3
Label: Verified=0