Update patch set 1
Patch Set 1:

I was hoping that keeping the Change-Id will merge this with the original submission https://git.opendaylight.org/gerrit/#/c/3968/ but I guess the submitter must be the same for that to happen. The author of the patch is still the same.

I think at this point it's best to wait for Colin to choose wether he wants to amend the original submission or just merge this one.

Sorry for the noise.

Patch-set: 1
Reviewer: Gerrit User 491 <491@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0