Update patch set 1
Patch Set 1:

Hi Sachi

can you please put this code in the release-1.0.X branch?
we use that branch for release and quick fixes for them. In the develop the fix will only be released in the next big release which won't be in the near future.

Patch-set: 1
Reviewer: Gerrit User 656 <656@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0