Update patch set 3
Patch Set 3:

Hi David,

In LocatorRecord if we keep as byte then it will accept values from -128 to 127 only, Any value greater than 127 will result in failure.And in Yang it is defined as uint8 whose value ranges from 0-255
Since they are supposed to be a number between 0-255, so I have converted it as short in LocatorRecord and used RangeConstraint method to put a check on range to accept the number between 0-255. Any value out of this range will result in IllegalArgumentException.

Patch-set: 3
Label: Verified=0