YANG changes to support SMR
Two main changes:

  1) Add support for sending Map-Request messages (with the SMR bit set)

  2) In all messages add support for both source and destionation IP
     addresses

Not sure on how to proceed on 2).  In this RFC I just renamed the
grouping TransportAddress to TransportAddresses and changed the leaves
to have source and destination IPs and ports.  This seemed like the
simplest change, and most futureproof, meaning that we may not need both
addresses in all messages now, but maybe we will in the future.  This
change would require some code changes in other places (not included in
this patch).

Another approach would be to just modify the MapRequestNotification
grouping and replace the "uses TransportAddress" with two
TransportAddress leaves, one for source, one for destination.  This way
we can get the source address of a Map-Request, and don't have to modify
so much code as in the first case.

And I can think of other combinations as well, but these two seeemed the
most obvious ones, would like to see what you think.

Change-Id: I6ee319c78d00318387abdfd1514b31701e354fa9
Signed-off-by: Lorand Jakab <lojakab@cisco.com>