Update patch set 2
Patch Set 2:

Looks good Lori, just one comment about the HashSet vs HashMap (For some reason the comment was not submitted last week when I wrote it). But now I think HashSet may be just fine for now. Specially that the main method we need is to iterate over the set members, and that will be less straight forward with HashMap. So lets keep it as it is. I can verify this in the setup once you submit the next revision with the fix for the integration test. Thanks

Patch-set: 2
Reviewer: Gerrit User 654 <654@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0