Add xTR-ID/Site-ID (de)serialization
The xTR-ID/Site-ID are not yet stored in the DAO, this is just support
to serialize/deserialize.  It will be used for the inventory service,
which will use the xTR-ID with the source RLOC as the internal xTR
identifyier.

This comment also cleans up the Map-Notify message, which doesn't have
the Proxy and Want-Map-Notify bits.

Change-Id: I37695e73f19822c28ca16f0eb53e028a1cf2651a
Signed-off-by: Lorand Jakab <lojakab@cisco.com>