Add mappingservice-netconf bundle
The goal is to enable the management of NETCONF connections to LISP
devices via a simple REST API. If the functionality turns out to be
useful to other controller services, a Java API may also be defineda in
the future.

The bundle consists of a netconf-connector data model that defines RPCs
for spinning up/down sal-netconf-connectors and the associated service
implementation. Both dependency resolution and netconf-connector
instantiation/tearing down are done programmatically via JMX.

Change-Id: I6db79956c1c43b57a6386366c9b4f16937aa53cc
Signed-off-by: Florin Coras <fcoras@ac.upc.edu>