Update patch set 2
Patch Set 2: Code-Review-1

I didn't look at the patch in detail yet, but I see you include autogenerated code in the yang-gen-sal and yang-yen-config directories.  I think we should add those directories to .gitignore, so that we avoid this in the future.  I will send a patch to do that.

Patch-set: 2
Label: Code-Review=-1