Update patch set 3
Abandoned

With all the changes for bumping the versions, I couldn't reliably rebase it, so I just downloaded and applied the latest version of this patch, applied Version-bump.sh to it, and created a diff from the current develop.  I wanted to upload the new changeset here, but the Change-Id is duplicate with 11702, and I couldn't find a way to specify which gerrit id the new patch set should go to.  So I created a the new Gerrit 11884

Patch-set: 3
Status: abandoned