Add coding style checks
Run the maven-checkstyle-plugin during the build, using rules from the
controller project, which include:

  - No TAB characters
  - No trailing spaces
  - No redundant, unused or star imports
  - Long constants defined with upper ell: 'L'
  - No empty statements (standalone ';' semicolon)
  - If a class overrides equals() it must override hashCode()

These rules are applied to .java, .xml, and .yang files.  Rule
violations are reported on the console, but they don't cause the build
to fail.  This behavior should be changed in the future, when the
existing code base is cleaned up, so that no new violations are
introduced after that.

Change-Id: I552eb72e1ac4b96a4cf80a216a1638bb36d77821
Signed-off-by: Lorand Jakab <lojakab@cisco.com>