Move Karaf distribution POM file to "best practices"
We introduce a new pom file with a <dependencyManagement> section
inclusing all lispflowmapping artifacts.  Since the "best practices" now
use different parent POMs for different parts of the build, this is to
avoid repeating the project artifacts in many files.

With this commit, our Karaf distribution has a very simplified POM file,
with the parent being managed in the controller project.  We only have
to specify our own dependencies.

Change-Id: Ibbe9e6baa4ce59c160cd704ccb2c3497bac89425
Signed-off-by: Lorand Jakab <lojakab@cisco.com>