Remove OSGi distribution build
The OSGi distribution was deprecated in Helium (we only shipped Karaf
features officially) and is being removed in Lithium.  The distribution
that is being built right now is based on
org.opendaylight.controller:distribution.opendaylight:0.2.0-SNAPSHOT,
which is Helium based, so we're building a Helium distribution with
lispflowmapping artifacts from Lithium.

Change-Id: I2ccf40e0626f3bdcf1d818fadca70c1bdfd5158d
Signed-off-by: Lorand Jakab <lojakab@cisco.com>