Bug 2871: Add support for IID mask-len
draft-farinacci-lisp-lcaf-07 introduced a change for LCAF type 2
(segmentation) where the 2nd reserved field in the LCAF header becomes
IID mask-len.  Some LISP implementations set this field to a non-zero
value when registering mappings.  LISPflowmapping needs to be able to
serialize-deserialize this field in order to authenticate such
registrations.

This commit adds support for the above mentioned ser/des, however, it
does not implement support for IID masking.  It doesn't fully fix bug
2871 either.

Change-Id: I777e5bb5bcabc4dc7b558ee6eac89d88e0820d8e
Signed-off-by: Lorand Jakab <lojakab@cisco.com>