Prepare the YANG models for RESTCONF
This commit is in preparation of an MD-SAL based mapping database.  The
existing YANG model had to be modified in order to be useble for that
goal, and a new model was added to describe the new API.  The models
have been checked to pass pyang.  The implementation of the MD-SAL based
mapping database will be provided by another commit.

Change-Id: Idb1a31a00fa110f2a1c493672cb508f23cf5fad7
Signed-off-by: Lorand Jakab <lojakab@cisco.com>