Enforce basic checkstyle on mappingservice.implementation
Odlparent checkstyle rules are too restrictive for us at this time, so
we're ignoring them.  Instead of doing that, this patch overrides the
odlparent checkstyle rules for the mappingservice.implementation bundle
which we modify frequently, so that we enforce at least some basic
rules.

The commit also fixes some checkstyle issues that were discovered.

Change-Id: Ife358094b206fe1f7b70fef25dfb43c9cea403ca
Signed-off-by: Lorand Jakab <lojakab@cisco.com>