Bug 3781: Return all locators on the RPC interface
When using the RPC interface to retrieve mappings, the DAO backend is
queried for the data.  Howver, the DAO doesn't store the 'name' list key
for a mapping, and so far the string "Autogenerated" was returned for
each locator.  It looks like the ODL RPC implementation doesn't allow
list entries with duplicate keys, so only one locator was returned to
the user (even though the LISP code handled all).

This commit changes to "Autogenerated" string to be the textual
representation of the locator.  This does not fully solve the bug,
since it is theoretically possible to list the same locator more than
once in a mapping.  That is not the case for normal operation however,
so at least we can lower the severity of the bug from 'critical' to
'minor'.

The proper fix is to store the user supplied key for the locator in the
DAO.

Change-Id: Icdfeb4a7ce9a79c16e854e8e92ebd6d1f7d30ed5
Signed-off-by: Lorand Jakab <lojakab@cisco.com>
(cherry picked from commit a547901fc0f71feb4216bda5d69a73c82947e679)