Add basic Site-ID support to mappings
This patch adds basic support for storing a list of Site-IDs with a
mapping in the MD-SAL config datastore, both from the southbound via
Map-Register messages and from the RESTCONF interface. Using RPCs is not
supported yet. This is useful for grouping EIDs belonging to the same
site in the topology view of the LISP UI.

Change-Id: Ie2ffc2dfb5f922ca5be3e3282ef63af0f0054f43
Signed-off-by: Lorand Jakab <lojakab@cisco.com>