Move SB RPCs from LISP Proto to API
The lisp-proto artifact contains the YANG model describing the LISP
protocol messages.  It also contained RPCs that are used to send those
protocol messages.  However, those RPCs are not part of the LISP
protocol itself, and this commit moves them to the API bundle.  They are
implemented the the southbound plugin.

Change-Id: I695068ba5a71db4fe3d5ae6560f03570bf9f3352
Signed-off-by: Lorand Jakab <lojakab@cisco.com>