Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Generally looks good. One minor comment/question in line. Aside from that, we should make sure these changes does not effect SFC. It shouldnt as we checked and they only rely on mappingservice and southbound features which we did not change. Nevertheless, can you run/install SFC fresh with this change to make sure we are not breaking it? thanks.

Patch-set: 2