Switch to using lisp-address instead of LispAddress
This commit modifies the build to make experimentation easy. Modules
that wouldn't build are commented out, tests that would fail are skipped
by default, and Karaf will only load odl-lispflowmapping-models instead
of the odl-lispflowmapping-msmr feature. Additionally odl-mdsal-apidoc
is pulled in automatically.

If you start Karaf after a successful build, you should be able to
explore the API and play with adding/removing keys/mappings to/from the
config datastore.

Change-Id: I70fd3c950d4b668bc57fa55a8156b76fe2751437
Signed-off-by: Lorand Jakab <lojakab@cisco.com>