Sonar: Objects should be compared with 'equals()'
Fix 9 instances of Sonar issue "Objects should be compared with
'equals()'" labeled as Major. The rule description says: "Using the
equality (==) and inequality (!=) operators to compare two objects does
not check to see if they have the same values. Rather it checks to see
if both object references point to exactly the same object in memory.
The vast majority of the time, this is not what you want to do. Use the
.equals() method to compare the values of two objects or to compare a
string object to a string literal."

Change-Id: I491981a10b8a95784d313b17ac23a693f9a8b685
Signed-off-by: Lorand Jakab <lojakab@cisco.com>