Add support for non-overwriting mapping storage
When the "lisp.mappingOverwrite" configuration variable is set to
"false" this new mode of operation will allow storing a separate
MappingRecord for each unique xTR-ID from southbound registration. This
is independent from (but in preparation of) the merge functionality, as
rataining all southbound registrations is useful in non-merge mode too,
i.e., for liveness detection and statistics.

Change-Id: I6d322fbb5620815dbfc09dc49770373b48ec91c2
Signed-off-by: Lorand Jakab <lojakab@cisco.com>