Add support for Service Path mapping records processing
The patch introduces two important changes for mapping records with
Service Path EIDs. First, it ensures that ServicePath EIDs are
normalized (i.e., Service Index is set to 0) before being used as keys
in mappings caches. Second, when doing lookups, for records with ELP
locator records, the Service Index in the requested EID is used to
select one, next hop, locator that is returned in the map-reply, instead
of the whole ELP.

Change-Id: Id487f3603f89beb5a878d5256709112ef004b4f6
Signed-off-by: Florin Coras <fcoras@cisco.com>