Update patch set 1
Patch Set 1: Code-Review-1

Didnt mean to +2 before. So as I mentioned in a separate email, I confirmed with Johnson that Map Notify messages in case of merge enabled includes the full list of merged locators. So to interoperate with IOS, we have to do the merge at store/write, and keep a fully merged locator set under Subkey RECORD. Lori, could you please Ack on this?
Thanks.

Patch-set: 1
Label: Code-Review=-1