Update patch set 1
Patch Set 1:

> Didnt mean to +2 before. So as I mentioned in a separate email, I
 > confirmed with Johnson that Map Notify messages in case of merge
 > enabled includes the full list of merged locators. So to
 > interoperate with IOS, we have to do the merge at store/write, and
 > keep a fully merged locator set under Subkey RECORD. Lori, could
 > you please Ack on this?
 > Thanks.

Well, I guess that settles it :) so ACK.

In that case, I would like to change course on "the SimpleMapCache is implementation agnostic" discussion. We can rename it to LispSouthboundMapCache to show that it's not generic anymore, but I would like to do the merging logic inside, and avoid asking current status from an upper layer like MappingSystem, and then sending back results, because it would complicate interfaces, and make performance optimizations harder. We already loose some performance by doing merge at registration, and at this point I think we should do it in the cache. We can discuss this further in today's call, but I wanted to raise the topic earlier, we all can weigh pros and cons before the call.

Patch-set: 1
Label: Verified=0