Bug 5229: Fix NPE in MappingRecordSerializer
When merge is on, the code tries to get the current merged mapping, and
send that in the Map-Notify message.  However, when a mpping is first
registered, the lookup result is null.  This commit fixes that, and uses
the mapping from the Map-Register packet in the Map-Notify when the
lookup is null.

Change-Id: I6f85d225071dc46e80f1453ae47abd037ebf4224
Signed-off-by: Lorand Jakab <lojakab@cisco.com>