Update patch set 2
Patch Set 2: Code-Review+2

When I updated the serializers for the IETF YANG model, I did this change too, but I faced some problems, and had to revert to the old way of doing things.  I don't remember the exact failures, but I couldn't reproduce anything today, so I'll go ahead and merge this. I just wanted to note that if some address related failures crop up, this is something that may be the cause.

Patch-set: 2
Label: Code-Review=+2