Update patch set 1
Patch Set 1:

(1 comment)

Jozef, I just realized some of the methods are not used within the code. Could you mark all such methods as deprecated? We should remove them in the near future. Note however that some are used in sfclisp so it would be good if you could clone the sfc code as well.

Patch-set: 1
Label: Code-Review=0