Update patch set 2
Patch Set 2:

(3 comments)

Jozef, although I recommended doing this, it may have not been the best of ideas. As far as I can tell, most of the methods are used in other bundles (so they shouldn't be deprecated) and those that are not, I wouldn't mark as deprecated for now. So I think it may be best to just abandon this.

Patch-set: 2
Reviewer: Gerrit User 1966 <1966@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0