Update patch set 7
Patch Set 7:

> Yes. But I still don't think we should remove 'javadoc:aggregate'
 > goal from JJB, since we link to out older APIs in the wiki, that
 > are in Jenkins. Not sure if I should backport this to older
 > branches?

As long as the builds are still working with the javadoc:aggregate goal then I guess it's okay to leave the goal in place. My only concern was if it caused conflict but since the merge job pass I assume everything is fine.

I think we should backport to Beryllium. Phil will be very pleased to see more projects have javadocs generated in Nexus for Beryllium builds.

Patch-set: 7