Checkstyle: enforce 120 character limit
We spend too much time on this character limit in reviews. Fix existing
violations and enforce the limit at build time in the future to avoid
this.

Not all fixed violations resulted in beautifully readable code, but this
is a first step, to get the enforcement off the ground, things can be
then improved gradually, as we work on the code in question.

Change-Id: I0d4a9aeca0e7a8f9f9927581211b40d13bcd9661
Signed-off-by: Lorand Jakab <lojakab@cisco.com>