Allow setting some configuration variables
The goal of this patch is to enable writing integration tests that can
change configuration variables, so that different setups can be tested.
It is not a complete patch, we still need to discuss what our approach
to configuration will be, but we have to at least unblock integration
test development.

Change-Id: Ib5c8909f6e85086370ea3533a3b03b373d7109c9
Signed-off-by: Lorand Jakab <lojakab@cisco.com>