Update patch set 14
Patch Set 14:

(4 comments)

Just a few comments inline, not a full review. I have some second thoughts about YANG modeling though, I have to think a bit more and discuss tomorrow, but it looks like we have a pseudo-YANG-model MapRegisterCache which was initially autogenerated and then used here, and then we have the mappingKeepAlive notification and we're copying between the two, which is not the most efficient. I'll get back to you on how to handle this better.

Patch-set: 14
Label: Verified=0