Adapt updateMappingRegistration() methods for cache
We have updateMappingRegistration() methods in our APIs, but they are
currently unused, and only take the EID as the input, to update its
mapping with the current system time as timestamp.

In preparation of the Map-Register cache, this patch updates the API and
code to use the EID and the MapRegisterCacheMetadata as arguments to the
method. While in first phase we only need the timestamp from the
metadata, when caching for merging will be implemented, we will need
more information (at least xTR-ID and merge bit). To avoid adding more
and more paramenters in the future, just use the
MapRegisterCacheMetadata object from the start.

Change-Id: Id1b0721cccd72b31470542a1ff5d6c4b16766e65
Signed-off-by: Lorand Jakab <lojakab@cisco.com>