Update patch set 1
Patch Set 1:

(4 comments)

> One alternative to passing the MapRegisterCacheMetadata (which is
 > really a south bound object, and I am not sure if passing it like
 > this all the way to Map-Cache is a good idea) all the way to
 > mapcache is to parse it in mapping-service/ system to go over the
 > list of EIDs, and, for merge case, create a separate method /API
 > for update with merge, which also gets the xTR-ID.  But I am not
 > strongly against the way it is done here, so if you think this is
 > better, lets stick to it. 

I changed this several times during implementation, and I was on the fence on which approach is better. I chose the metadata, since it's easier to change that when we need additional info than changing APIs, or keep adding new APIs.

However, since I was undecided and you prefer the non-metadata approach, I will switch to that.

> But still the list of EIDs in metadata
 > needs to be parsed in MappingService, and updateMappingRegistration
 > should be called per EID.

Patch-set: 1
Label: Verified=0