Update patch set 2
Patch Set 2:

Hi Lori,
Do you think we need the timestamp to be passed from SB? the existing method uses the current time-stamp, and I am thinking that is fine, and there is no need to carry the timestamp form SB for this update. Basically, I think for the non-merge case we dont really need to change the mapcache updateMappingRegistration method. What we do need to do is to iterate over the EID list in metadata in LISPMappingService, and call the updateMappingRegistration from there per EID. Then the current timestamp is set per EID in the mapcache. right? Again, not strongly against passing the timeStamp if you think it is necessary, but if not, we should keep the method as is in the mapcache.

Patch-set: 2