Spin out mapcache package from implementation
We will need SimpleMapCache in both implementation and southbound, hence
splitting out the mapcache package from implementation as its own
bundle. As a result MappingMergeUtil needs to be split into two part,
with a small overlap. Some config values are rewritten as constants
until a proper configuration infrastructure is in place, or we at least
spin out the config package too.

Change-Id: Iff6aaeb79a2e310e62c343aa2ca6d5fc868e5a94
Signed-off-by: Lorand Jakab <lojakab@cisco.com>