Update patch set 3
Patch Set 3:

(5 comments)

I dont follow why widest negative prefix is not needed for multiTableMapCache. I think it should be done at least the same as Simple Map Cache. On top of it, we need to decide how to address explicit negative mappings added from NB. I am ok with making that one a separate patch (need to build trello card for it to remember) in order to get this one in sooner. But as is, I think the patch will result in incorrect negative mappings when NB mappings exist.

Patch-set: 3
Label: Verified=0