Update patch set 2
Patch Set 2:

The project is now using blueprint, so to inject the EntityOwnershipService you should add a reference to it in the blueprint.xml file.

Reference to add would be something like this: https://github.com/opendaylight/openflowplugin/blob/master/openflowplugin-blueprint-config/src/main/resources/org/opendaylight/blueprint/openflowplugin.xml#l10

File where to add would be this one:
https://github.com/opendaylight/lispflowmapping/blob/master/mappingservice/implementation/src/main/resources/org/opendaylight/blueprint/mappingservice.xml

Then you can inject the dependency in a class as an argument in the constructor.

Patch-set: 2