Adding authentication data to map register cache.
Authentication type and string are stored in map register cache in map value.
If map register cache hit occures and it is required to send map notify
message which should be authenticated then this localy stored authentication
data are used.

If authentication data are changed (value in datastore) then the application is
switched to authentication refreshing mode for period of time which is equal
to map register cache record timout. During this period are authentication data
obtained as before - by direct request to simple map cache.

Change-Id: I6537b8cd88773c960a3f4d90b8e26ea2bb397fb0
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>