Update patch set 1
Patch Set 1:

Thanks for the "no testing mark" Jozef. 
Code looks good, and as Lori described, the idea was to use the Auth Key stored in the cache (when there is a hit) for Map Notify Auth Data , instead of looking up in smc (because of longest prefix matching, the lookup may turn into 32 lookups for IPv4 or 128 lookups for IPv6). 
@Lori: I just noticed a problem though, what happens if the AuthKey is changed via NB? we wont be able to notice that! So either we have to limit the amount of time (or number of times) we can keep something in the cache (e.g. up to 5 hits per Map Register) or we have to properly remove cache entries if their Auth Key is changed. I think the latter is going to be tough, because multiple EIDs can be in a single Map Register, and only the Auth Key for one of them can change. So I am tempted towards the first one (keep a hit counter per cache entry, and invalid a cache entry after  x number of hits). Thoughts?

Patch-set: 1
Label: Code-Review=0