Checkstyle: prepration for switching to odlparent
There has been a lot of work on the ODL-wide checkstyle rules in
odlparent, and we should eventually be compliant. To be able to make
progress on individial bundles, this commit moves our custom checkstyle
enabler from mappingservice/pom.xml to the individual child poms. As an
example, mappingservice/shell is updated to comply with ODL checkstyle
rules and turns on enforcement. Additionally, it contains a violation
that is inevitable, and code to suppress than one particular violation.
That can be used as an example in other places where there is good
reason not to follow ODL style.

Change-Id: I3455f4084fa1302f8a970b2dfba0b9293767aecf
Signed-off-by: Lorand Jakab <lojakab@cisco.com>