Update patch set 2
Patch Set 2:

(1 comment)

General question: from our discussion in the last meeting my feeling was that the singleton service would only allow one cluster node to be fully up at a time. So we said that we keep the old implementation as well, so that we can have passive replication on slaves. Looking at this patch I have the feeling that we are actually keeping the old behavior, just using different APIs. Is that correct?

Patch-set: 2