Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/lispflowmapping-verify-beryllium-mvn33-openjdk8/1/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/lispflowmapping-verify-beryllium-mvn33-openjdk8/1/ )

https://jenkins.opendaylight.org/releng/job/lispflowmapping-verify-beryllium-mvn33-openjdk7/1/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/lispflowmapping-verify-beryllium-mvn33-openjdk7/1/ )

https://jenkins.opendaylight.org/releng/job/lispflowmapping-validate-autorelease-beryllium/146/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/lispflowmapping-validate-autorelease-beryllium/146/ )

https://jenkins.opendaylight.org/releng/job/lispflowmapping-distribution-check-beryllium/71/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/lispflowmapping-distribution-check-beryllium/71/ )

Patch-set: 1