Update patch set 1
Patch Set 1:

well, we can still use the wrapper for passing data back and forth betw simpleMapCahce and MappingMergeUtil. The main concern I have is the storage in DAO. Basically with this patch we are storing MappingRecord in the EID DAO, but the wrapper in the xtrId DAO. Since so far we assumed the value objects are identical, I wonder if this change would have side effects. Basically, I am fine with the wrapper approach in general,  the question is should we try to keep the value object in the two DAOs instances identical? The main side effect I see with storing the wrapper in both DAO's is the extra object creation at time of addMapping. Other than that I agree that having the wrapper is more readable than separate sub records for date and record. But then again, we cant get rid of sub-records all together, because of src/dst and subscribers.

Patch-set: 1
Label: Verified=0