Update patch set 1
Patch Set 1:

I agree we can't get rid of sub-records completely, and that's fine. That said, I think we should still minimize their use.

The only code saving MappingRecords for xTR-IDs is SimpleMapCache, and that subDAO is only accessible there. So I think it's ok to store a different Object. But I agree it is confusing to have it under a sub-key with the same name (RECORD in this case), so I think the xTR-ID subDAO should then store ExterndedMappingRecords under a different sub-key, like EXT_RECORD or XTR_ID_RECORD. Would that be a good solution to your concerns?

Patch-set: 1