Allow setting SB registration validity value
For integration testing it helps speed things up if we don't have to
wait the full 200 seconds to elapse before a southbound entry expires
(which is minimum allowed value to be set). This patch allows setting
values lower than 200 s as the southbound timeout, and logs a warning
when that happens.

Change-Id: I9532d74d85c5f519e09480e92c1f8f542fefcc93
Signed-off-by: Miroslav Toth <mirtoth@cisco.com>
Signed-off-by: Lorand Jakab <lojakab@cisco.com>