Update patch set 2
Patch Set 2:

Hi Shakib, The course timer bucket feature is not really associated with this bug number. Can you please update the title and remove the bug pointer, it is a bit confusing. No bug pointer is needed for this patch. Just mention bucket  timer for SB mappings of something of that sort. Thanks.

Patch-set: 2
Label: Verified=0