WIP: Make SimpleMapCache simple again
A lot of logic in SimpleMapCache is intelligence to handle LISP specific
features, like merging mappings from several different xTRs and handling
expiration of mapping state from UDP xTR registrations. This logic does
not belong there architecturally, and should be moved out.

SimpleMapCache still needs to handle some LISP specific functionality,
but methods implementing that functionality should not be part of the
general IMapCache contract, which is implemented by all map caches. This
commit introduces a new interface, ILispMapCache, which is only
implemented by SimpleMapCache, to provide access to xTR-ID specific
mappings, and timestamp updates.

Since FlatMapCache was not really maitained and tested, this patch
causes the map caches not be built when `iterateMask` is false. It's
better to fail than give undefined behavior to the user.

The patch also updates the map-cache, mapping system and mapping service
interfaces to use ExtenededMappingRecord instead of the immutable
MappingRecord (contained within) so that mapping context can be easily
carried in both directions between layers.

WIP: This patch set can be built, but only with `mvn clean install
-Dmaven.test.skip=true`, as the tests are not yet updated (and don't
build as of now).

Change-Id: I49dbc68615a360d1b9d1c32b4af23726f173259f
Signed-off-by: Lorand Jakab <lojakab@cisco.com>