SMR parent prefix continued
This patch completes work started in
https://git.opendaylight.org/gerrit/#/c/47120/

First, it handles the insertion of a positive mapping that overlaps
a negative one, by removing the negative mapping and generating SMRs for
its subscribers. This leads to pushing the positive mapping to the right
subscribers and generating only the needed negative mappings.

Second, it merges negative mappings when a positive mapping is removed
(reversing what was done in the first point above).

Change-Id: I6a8799d3af23c20a9797840b3e64f632a3e2ee91
Signed-off-by: Lorand Jakab <lojakab@cisco.com>