Bug 7272: Fix prefix removal in MultiTableMapCache
This fix is a bit of a brute-force, since it removes authentication keys
and all data, when removing a mapping, but for the MultiTableMapCache in
lispflowmapping as-is currently it is OK, since authentication keys are
stored in SimpleMapCache. So this is just a band-aid for the particular
use case reported in the bug, but a proper fix will require more
changes, which need to be discussed.

Change-Id: I295d2921f6b81505271887fbde2078ec38dbca0d
Signed-off-by: Lorand Jakab <lojakab@cisco.com>
(cherry picked from commit 0241b00bf9dddb826ae71904341de30e7f0b19a1)