WIP: Cleaning operational should clean map-cache too
The recently added remove-all-operational-content RPC removes the
operational datastore content, but the in-memory southbound map-cache is
not cleaned on master nodes, since southbound operations go to the cache
first. This patch explicitly clears the southbound map-cache on master
nodes when the RPC is called.

It is still WIP, because it needs another patch (under discussion) to
separate authentications keys into a different SimpleMapCache. Right now
calling the RPC would remove authentication keys too, but they are in
the config datastore, so that would be incorrect behavior.

Change-Id: I3c1b899de592f5499779c55fb880481a60f9a64a
Signed-off-by: Lorand Jakab <lojakab@cisco.com>