Make 'mappings' CLI output user friendly [WIP]
Before this patch, the mappings CLI command would print out the full
Java objects contained in a map-cache. That's great for in-depth
debugging, but for a quick look it very hard to read. This patch
implements a user friendly CLI output for mappings, to improve
readability.

Right now the old full output is no longer available in Karaf, but is
planned, by adding a '-d' (from debug) switch to the CLI commands.

Example output:

---8<--------------------------------------------------------------------
opendaylight-user@root>mappings
Policy map-cache
----------------

Southbound map-cache
--------------------
Instance ID 0
  192.0.2.1/32, TTL: 1440
    -> Locator                                         State     Pri/Wgt
       192.168.16.21                                   up        1/1
       fdab:cc19:b80e::21                              up        1/1
       -----------------------------------------------------------------
    -> Subscriber RLOC                                 Subscriber EID
       192.168.16.21                                   No Address Present
       127.0.0.2                                       No Address Present
-------------------------------------------------------------------->8---

Change-Id: Id3c304ac9eda77a0cc36bff1cdeb18d61608a796
Signed-off-by: Lorand Jakab <lojakab@cisco.com>