Add toString() to MappingData
When running the `mappings` command on the Karaf console to view current
in-memory data structures, the address field of mapping shows the memory
reference of the MappingData object, since it doesn't override the
toString() method. This happens since the SimpleMapCache class was
simplified in https://git.opendaylight.org/gerrit/#/c/46778/ and the
MappingRecord was wrapped into a container class.

This patch implements the toString() method for the MappingData class.

Change-Id: Ib39ddb2c6330c28734f948d854ed42925cb1aa3a
Signed-off-by: Lorand Jakab <lojakab@cisco.com>