Bug 7628: Provide Karaf 4 features [WIP]
While SingleFeatureTest is passing for many of the Karaf 4 features with
'mvn test -Psft', the only feature that can be loaded in the generated
Karaf 4 distribution is odl-lispflowmapping-models. There is a known
build issue with the aaa projects, which is needed by odl-restconf, and
also some strange issue with odl-mdsal-broker.

Change-Id: Icf8d404f4d46c71557b1f6f088626c2fb75200c8
Signed-off-by: Lorand Jakab <lojakab@cisco.com>