Provide Karaf 4 features
This patch provides Karaf 4 features for lispflowmapping. The dlux and
neutron upstreams don't have their Karaf 4 patches finalized and merged
yet, so our odl-lispflowmapping-neutron and odl-lispflowmapping-ui
features for Karaf 4 are not enabled yet (but they are also migrated and
ready to be enabled).

The only Karaf 4 feature that work properly for now in the Karaf 4
distribution is odl-lispflowmapping-models. Other features can be
loaded, but the console if full of config subsystem (CSS) errors and the
feature functionality is not working. This is a know upstream issue that
is being worked on and needs no fix in lispflowmapping.

Once the CSS issues are ironed out, we can start working on migrating
our integration tests to a Karaf 4 based container.

Bug: 7628
Change-Id: Icf8d404f4d46c71557b1f6f088626c2fb75200c8
Signed-off-by: Lorand Jakab <lojakab@cisco.com>